revert back itemsToComboboxItems
