+++ src/components/Navbar/SettingsDrawer/TranslationSection.tsx
@@ -21,8 +21,23 @@ import { numbersToStringsArray, stringsToNumbersArray } from 'src/utils/array';
 import { throwIfError } from 'src/utils/error';
 import useSWR from 'swr';
 import { makeTranslationsUrl } from 'src/utils/apiPaths';
-import { itemsToComboboxItems } from 'src/utils/input';
+import { getTranslatedLabelWithLanguage } from 'src/utils/input';
+import { DropdownItem } from 'src/components/dls/Forms/Combobox/ComboboxItem';
+import AvailableTranslation from 'types/AvailableTranslation';
 import Section from './Section';
+
+// convert translations data (from API) to combobox items
+// so we can use Combobox component
+const translationsToComboboxItems = (translations: AvailableTranslation[]): DropdownItem[] =>
+  translations.map((item) => {
+    const stringId = item.id.toString();
+    return {
+      id: stringId,
+      value: stringId,
+      label: getTranslatedLabelWithLanguage(item),
+      name: stringId,
+    };
+  });
 
 const TranslationSection = () => {
   const dispatch = useDispatch();
@@ -47,7 +62,7 @@ const TranslationSection = () => {
     return null;
   }
 
-  const items = itemsToComboboxItems(translations);
+  const items = translationsToComboboxItems(translations);
 
   return (
     <Section>
